Класс app в отдельный файл
